+++ src/components/headless/components/data/item-scroller/item-scroller-page.props.ts
@@ -1,0 +1,10 @@
+import { TypeOrArray } from '@beesoft/common';
+import { ReactNode } from 'react';
+
+export interface ItemScrollerPageProps {
+  intersectionObserver?: IntersectionObserver;
+  resizeObserver?: ResizeObserver;
+  page: number;
+  height?: number;
+  children: TypeOrArray<ReactNode>;
+}