+++ src/index.ts
@@ -1,26 +1,35 @@
 // components
 import { BeeSoftProvider } from './common/contexts/beesoft.context.tsx';
-import { FormGroupItemOrientation, GroupChangeEvent } from './components/form/form-generic.interfaces.ts';
+import {
+  SelectionLabelLocation,
+  FormGroupItemOrientation,
+  GroupChangeEvent,
+} from './components/form/form-generic.interfaces.ts';
 import { Button } from './components/navigation/buttons/button/button.component.tsx';
 import { Checkbox } from './components/form/checkboxes/checkbox/checkbox.component.tsx';
 import { CheckboxGroup } from './components/form/checkboxes/checkbox-group/checkbox-group.component.tsx';
 import ContentEditableInput from './components/form/inputs/content-editable-input/content-editable-input.component';
 import { DateTime } from './components/form/date-time/date-time.component';
 import FormattedInput from './components/form/inputs/formatted-input/formatted-input.component';
+import { GroupButton } from './components/form/buttons/group-button/group-button.component.tsx';
 import { MediaQuery } from './components/mobile/media-query/media-query.component.tsx';
 import { MobileOverlayPanel } from './components/mobile/overlay/mobile-overlay-panel.component.tsx';
 import OverlayPanel from './components/overlay/overlay-panel/overlay-panel.component';
+import { RadioButton } from './components/form/radio-buttons/radio-button/radio-button.component.tsx';
+import { Toggle } from './components/form/checkboxes/toggle/toggle.component.tsx';
 
 // functions
 import { createBeeSoftTheme, applyBeeSoftTheme } from './components/common-functions';
 
 // types
 import { CheckboxChangeEvent } from './components/form/checkboxes/checkboxes.interfaces.ts';
-import { CheckboxLabelLocation, CheckboxRef } from './components/form/checkboxes/checkbox/checkbox.props.ts';
+import { CheckboxRef } from './components/form/checkboxes/checkbox/checkbox.props.ts';
 import { DateFormatType, DateSelectionType } from './components/form/date-time/date-time-types';
 import { DateTimeInputTemplateProps } from './components/form/date-time/date-time.props.ts';
 import { DateTimeScrollerTemplateProps } from './components/form/date-time/date-time-scroller.component';
 import { DateTimeCalendarTemplateProps } from './components/form/date-time/date-time-calendar.component';
+import { GroupButtonItemTemplateProps } from './components/form/buttons/group-button/group-button.props.ts';
+import { RadioChangeEvent } from './components/form/radio-buttons/radio-button/radio-button.props.ts';
 
 // styles
 import './index.css';
@@ -33,9 +42,12 @@ export {
   ContentEditableInput,
   DateTime,
   FormattedInput,
+  GroupButton,
   MediaQuery,
   MobileOverlayPanel,
   OverlayPanel,
+  RadioButton,
+  Toggle,
 
   // functions
   applyBeeSoftTheme,
@@ -43,7 +55,7 @@ export {
 
   // types to be used as values
   FormGroupItemOrientation,
-  CheckboxLabelLocation,
+  SelectionLabelLocation,
   DateFormatType,
   DateSelectionType,
 };
@@ -55,4 +67,6 @@ export type {
   DateTimeInputTemplateProps,
   DateTimeScrollerTemplateProps,
   DateTimeCalendarTemplateProps,
+  GroupButtonItemTemplateProps,
+  RadioChangeEvent,
 };